+++ packages/component-library/src/components/FormMarketoEmbed/index.ts
@@ -1,0 +1,2 @@
+export { default } from './FormMarketoEmbed';
+export type { FormMarketoEmbedProps } from './FormMarketoEmbed';