+++ packages/component-library/src/index.tsx
@@ -22,3 +22,4 @@ export { default as NavigationBar } from './components/NavigationBar';
 export { default as mockTheme } from './theme/mock.theme';
 export { default as createTheme } from './theme/createTheme';
 export { default as BackToTop } from './components/BackToTop';
+export { default as FormMarketoEmbed } from './components/FormMarketoEmbed';